requirements.txt
Flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.5.0
google-auth>=2.30.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pyodbc>=4.0.39
ddgs>=1.0.0
tqdm>=4.66.0

